# inflate: Generate code for byte readers