… flate/gen_inflate.go
… flate/inflate_gen.go