More than 20% additional speedup:

```
λ benchcmp old.txt new.txt
benchmark                       old ns/op     new ns/op     delta
BenchmarkGunzipCopy-32          19698008      16125274      -18.14%
BenchmarkGunzipNoWriteTo-32     19883807      16071205      -19.17%
BenchmarkGunzipStdlib-32        23397455      23201339      -0.84%

benchmark                       old MB/s     new MB/s     speedup
BenchmarkGunzipCopy-32          242.31       295.99       1.22x
BenchmarkGunzipNoWriteTo-32     240.04       296.99       1.24x
BenchmarkGunzipStdlib-32        203.99       205.72       1.01x
```