@klauspost 2020-02-24
Thanks to @dsnet for pushing me into this :)